refactor(router): event/sponsor/account のルート定義を分割
