+++ apps/dashboard/lib/core/router/account.dart
@@ -1,0 +1,11 @@
+part of 'router.dart';
+
+class AccountBranch extends StatefulShellBranchData {}
+
+class AccountInfoRoute extends GoRouteData {
+  const AccountInfoRoute();
+
+  @override
+  Widget build(BuildContext context, GoRouterState state) =>
+      const AccountInfoScreen();
+}
